refactor: extract ChatView composer command menu
Move ComposerCommandMenu, ComposerCommandMenuItem, ComposerCommandItem, and the shared VS Code entry icon helper into a dedicated component file. Keep ChatView responsible for state and event wiring while reducing the size of the main component.
